+++ packages/datasource-mongoose/test/integration/many-to-many/collection_list.test.ts
@@ -5,7 +5,7 @@ import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factorie
 import { Connection, Types } from 'mongoose';
 
 import MongooseDatasource from '../../../src/datasource';
-import { setupWith2ManyToManyRelations } from '../../_helpers';
+import { setupWith2ManyToManyRelations } from '../_helpers';
 
 describe('MongooseCollection', () => {
   let connection: Connection;
