+++ packages/datasource-mongoose/test/integration/many-to-one/collection_aggregate.test.ts
@@ -5,7 +5,7 @@ import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factorie
 import { Connection, Types } from 'mongoose';
 
 import MongooseDatasource from '../../../src/datasource';
-import { setupWithManyToOneRelation } from '../../_helpers';
+import { setupWithManyToOneRelation } from '../_helpers';
 
 describe('MongooseCollection', () => {
   let connection: Connection;
