+++ packages/datasource-mongoose/test/integration/review/collection_create.test.ts
@@ -5,7 +5,7 @@ import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factorie
 import { Connection } from 'mongoose';
 
 import MongooseDatasource from '../../../src/datasource';
-import { setupReview } from '../../_helpers';
+import { setupReview } from '../_helpers';
 
 const caller = factories.caller.build();
